Copy TypeScript's stdlib only if installed

+++ _setup_support.py
@@ -368,8 +368,9 @@ def install_js():
 
     if exists(target_tslibdir):
         shutil.rmtree(target_tslibdir)
-    lib = {"lib.es5.d.ts", "lib.dom.d.ts", "lib.es2015.core.d.ts", "lib.es2015.promise.d.ts"}
-    shutil.copytree(TSLIB, target_tslibdir, ignore=lambda _, files: [ f for f in files if f not in lib ])
+    if exists(TSLIB):
+        lib = {"lib.es5.d.ts", "lib.dom.d.ts", "lib.es2015.core.d.ts", "lib.es2015.promise.d.ts"}
+        shutil.copytree(TSLIB, target_tslibdir, ignore=lambda _, files: [ f for f in files if f not in lib ])
 
 # -----------------------------------------------------------------------------
 # Helpers for collecting package data
